+++ silk/tests/conftest.py
@@ -1,6 +1,8 @@
 # (C) Datadog, Inc. 2021-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+from copy import deepcopy
+
 import pytest
 
 from datadog_checks.dev import docker_run, get_here
@@ -23,4 +25,4 @@ def dd_environment():
 
 @pytest.fixture
 def instance():
-    return common.INSTANCE
+    return deepcopy(common.INSTANCE)
