+++ silk/tests/docker/silk_metrics/api/v2/stats/volumes_rw
@@ -1,114 +1,117 @@
-[
-    {
-        "iops_avg": 19011,
-        "latency_inner": 0.42999999999999999,
-        "latency_outer": 0.02,
-        "peer_k2_name": "K2-6310",
-        "resolution": 5,
-        "rw": "r",
-        "throughput_avg": 311469670,
-        "timestamp": 1644878230,
-        "volume": {
-            "ref": "/volumes/5"
+{
+    "hits": [
+        {
+            "iops_avg": 19011,
+            "latency_inner": 0.42999999999999999,
+            "latency_outer": 0.02,
+            "peer_k2_name": "K2-6310",
+            "resolution": 5,
+            "rw": "r",
+            "throughput_avg": 311469670,
+            "timestamp": 1644878230,
+            "volume": {
+                "ref": "/volumes/5"
+            },
+            "volume_name": "centos-vol1"
         },
-        "volume_name": "centos-vol1"
-    },
-    {
-        "iops_avg": 2138,
-        "latency_inner": 0.17999999999999999,
-        "latency_outer": 0.11,
-        "peer_k2_name": "K2-6310",
-        "resolution": 5,
-        "rw": "w",
-        "throughput_avg": 35035546,
-        "timestamp": 1644878230,
-        "volume": {
-            "ref": "/volumes/5"
+        {
+            "iops_avg": 2138,
+            "latency_inner": 0.17999999999999999,
+            "latency_outer": 0.11,
+            "peer_k2_name": "K2-6310",
+            "resolution": 5,
+            "rw": "w",
+            "throughput_avg": 35035546,
+            "timestamp": 1644878230,
+            "volume": {
+                "ref": "/volumes/5"
+            },
+            "volume_name": "centos-vol1"
         },
-        "volume_name": "centos-vol1"
-    },
-    {
-        "iops_avg": 2114,
-        "latency_inner": 0.17999999999999999,
-        "latency_outer": 0.11,
-        "peer_k2_name": "K2-6310",
-        "resolution": 5,
-        "rw": "w",
-        "throughput_avg": 34629222,
-        "timestamp": 1644878230,
-        "volume": {
-            "ref": "/volumes/7"
+        {
+            "iops_avg": 2114,
+            "latency_inner": 0.17999999999999999,
+            "latency_outer": 0.11,
+            "peer_k2_name": "K2-6310",
+            "resolution": 5,
+            "rw": "w",
+            "throughput_avg": 34629222,
+            "timestamp": 1644878230,
+            "volume": {
+                "ref": "/volumes/7"
+            },
+            "volume_name": "centos-vol3"
         },
-        "volume_name": "centos-vol3"
-    },
-    {
-        "iops_avg": 19024,
-        "latency_inner": 0.42999999999999999,
-        "latency_outer": 0.02,
-        "peer_k2_name": "K2-6310",
-        "resolution": 5,
-        "rw": "r",
-        "throughput_avg": 311695770,
-        "timestamp": 1644878230,
-        "volume": {
-            "ref": "/volumes/7"
+        {
+            "iops_avg": 19024,
+            "latency_inner": 0.42999999999999999,
+            "latency_outer": 0.02,
+            "peer_k2_name": "K2-6310",
+            "resolution": 5,
+            "rw": "r",
+            "throughput_avg": 311695770,
+            "timestamp": 1644878230,
+            "volume": {
+                "ref": "/volumes/7"
+            },
+            "volume_name": "centos-vol3"
         },
-        "volume_name": "centos-vol3"
-    },
-    {
-        "iops_avg": 2144,
-        "latency_inner": 0.17999999999999999,
-        "latency_outer": 0.11,
-        "peer_k2_name": "K2-6310",
-        "resolution": 5,
-        "rw": "w",
-        "throughput_avg": 35120742,
-        "timestamp": 1644878230,
-        "volume": {
-            "ref": "/volumes/6"
+        {
+            "iops_avg": 2144,
+            "latency_inner": 0.17999999999999999,
+            "latency_outer": 0.11,
+            "peer_k2_name": "K2-6310",
+            "resolution": 5,
+            "rw": "w",
+            "throughput_avg": 35120742,
+            "timestamp": 1644878230,
+            "volume": {
+                "ref": "/volumes/6"
+            },
+            "volume_name": "centos-vol2"
         },
-        "volume_name": "centos-vol2"
-    },
-    {
-        "iops_avg": 19198,
-        "latency_inner": 0.42999999999999999,
-        "latency_outer": 0.02,
-        "peer_k2_name": "K2-6310",
-        "resolution": 5,
-        "rw": "r",
-        "throughput_avg": 314543309,
-        "timestamp": 1644878230,
-        "volume": {
-            "ref": "/volumes/6"
+        {
+            "iops_avg": 19198,
+            "latency_inner": 0.42999999999999999,
+            "latency_outer": 0.02,
+            "peer_k2_name": "K2-6310",
+            "resolution": 5,
+            "rw": "r",
+            "throughput_avg": 314543309,
+            "timestamp": 1644878230,
+            "volume": {
+                "ref": "/volumes/6"
+            },
+            "volume_name": "centos-vol2"
         },
-        "volume_name": "centos-vol2"
-    },
-    {
-        "iops_avg": 19177,
-        "latency_inner": 0.42999999999999999,
-        "latency_outer": 0.02,
-        "peer_k2_name": "K2-6310",
-        "resolution": 5,
-        "rw": "r",
-        "throughput_avg": 314195968,
-        "timestamp": 1644878230,
-        "volume": {
-            "ref": "/volumes/8"
+        {
+            "iops_avg": 19177,
+            "latency_inner": 0.42999999999999999,
+            "latency_outer": 0.02,
+            "peer_k2_name": "K2-6310",
+            "resolution": 5,
+            "rw": "r",
+            "throughput_avg": 314195968,
+            "timestamp": 1644878230,
+            "volume": {
+                "ref": "/volumes/8"
+            },
+            "volume_name": "centos-vol4"
         },
-        "volume_name": "centos-vol4"
-    },
-    {
-        "iops_avg": 2143,
-        "latency_inner": 0.19,
-        "latency_outer": 0.11,
-        "peer_k2_name": "K2-6310",
-        "resolution": 5,
-        "rw": "w",
-        "throughput_avg": 35104358,
-        "timestamp": 1644878230,
-        "volume": {
-            "ref": "/volumes/8"
-        },
-        "volume_name": "centos-vol4"
-    }
-]
+        {
+            "iops_avg": 2143,
+            "latency_inner": 0.19,
+            "latency_outer": 0.11,
+            "peer_k2_name": "K2-6310",
+            "resolution": 5,
+            "rw": "w",
+            "throughput_avg": 35104358,
+            "timestamp": 1644878230,
+            "volume": {
+                "ref": "/volumes/8"
+            },
+            "volume_name": "centos-vol4"
+        }
+    ],
+    "total": 7
+}
